move types and type constructors from hm_algo into lambda_types
